+++ src/main/java/com/vagivagi/connector/toggl/wrapper/TogglWrapperNotExistProjectException.java
@@ -1,4 +1,4 @@
-package com.vagivagi.connector.toggl;
+package com.vagivagi.connector.toggl.wrapper;
 
 public class TogglWrapperNotExistProjectException extends RuntimeException {
     public TogglWrapperNotExistProjectException(String message) {
